Add navigation tabs for blockchain ecosystems supported by Keplr
