+++ src/App.tsx
@@ -1,144 +1,17 @@
-import React, { useEffect } from "react";
-import { OsmosisChainInfo } from "./constants";
-import { Balances } from "./types/balance";
-import { Dec, DecUtils } from "@keplr-wallet/unit";
-import { isAddress } from "@ethersproject/address";
-import { sendMsgs } from "./util/sendMsgs";
-import { api } from "./util/api";
-import { simulateMsgs } from "./util/simulateMsgs";
-import { MsgSend } from "./proto-types-gen/src/cosmos/bank/v1beta1/tx";
+import React from "react";
+
+import { OsmosisTab } from "./components/OsmosisTab";
+import { EvmTab } from "./components/EvmTab";
+import { StarknetTab } from "./components/StarknetTab";
+import { BitcoinTab } from "./components/BitcoinTab";
 import "./styles/container.css";
 import "./styles/button.css";
 import "./styles/item.css";
 
+type TabType = "osmosis" | "evm" | "starknet" | "bitcoin";
+
 function App() {
-  const [address, setAddress] = React.useState<string>("");
-  const [balance, setBalance] = React.useState<string>("");
-  const [recipient, setRecipient] = React.useState<string>("");
-  const [amount, setAmount] = React.useState<string>("");
-
-  const [evmResult1, setEvmResult1] = React.useState<string>("");
-
-  const [evmResult2, setEvmResult2] = React.useState<string>("");
-
-  const [evmResult3, setEvmResult3] = React.useState<string>("");
-
-  const [evmRecipient, setEvmRecipient] = React.useState<string>("");
-  const [evmAmount, setEvmAmount] = React.useState<string>("");
-  const [evmResult4, setEvmResult4] = React.useState<string>("");
-
-  const [evmChainId, setEvmChainId] = React.useState<string>("");
-
-  const [evmTokenAddress, setEvmTokenAddress] = React.useState<string>("");
-
-  const [keplrEip6963ProviderInfo, setKeplrEip6963ProviderInfo] =
-    React.useState<any>();
-
-  useEffect(() => {
-    window.addEventListener("eip6963:announceProvider", (e) => {
-      const event = e as CustomEvent;
-      if (event.detail.info.rdns === "app.keplr") {
-        setKeplrEip6963ProviderInfo(event.detail.info);
-      }
-    });
-
-    window.dispatchEvent(new Event("eip6963:requestProvider"));
-    init();
-  }, []);
-
-  const init = async () => {
-    const keplr = window.keplr;
-    if (keplr) {
-      try {
-        await keplr.experimentalSuggestChain(OsmosisChainInfo);
-        if (!keplr.ethereum?.isConnected()) {
-          await keplr.ethereum?.enable();
-        }
-      } catch (e) {
-        if (e instanceof Error) {
-          console.log(e.message);
-        }
-      }
-    }
-  };
-
-  const getKeyFromKeplr = async () => {
-    const key = await window.keplr?.getKey(OsmosisChainInfo.chainId);
-    if (key) {
-      setAddress(key.bech32Address);
-    }
-  };
-
-  const getBalance = async () => {
-    const key = await window.keplr?.getKey(OsmosisChainInfo.chainId);
-
-    if (key) {
-      const uri = `${OsmosisChainInfo.rest}/cosmos/bank/v1beta1/balances/${key.bech32Address}?pagination.limit=1000`;
-
-      const data = await api<Balances>(uri);
-      const balance = data.balances.find(
-        (balance) => balance.denom === "uosmo"
-      );
-      const osmoDecimal = OsmosisChainInfo.currencies.find(
-        (currency) => currency.coinMinimalDenom === "uosmo"
-      )?.coinDecimals;
-
-      if (balance) {
-        const amount = new Dec(balance.amount, osmoDecimal);
-        setBalance(`${amount.toString(osmoDecimal)} OSMO`);
-      } else {
-        setBalance(`0 OSMO`);
-      }
-    }
-  };
-
-  const sendBalance = async () => {
-    if (window.keplr) {
-      const key = await window.keplr?.getKey(OsmosisChainInfo.chainId);
-      const protoMsgs = {
-        typeUrl: "/cosmos.bank.v1beta1.MsgSend",
-        value: MsgSend.encode({
-          fromAddress: key.bech32Address,
-          toAddress: recipient,
-          amount: [
-            {
-              denom: "uosmo",
-              amount: DecUtils.getTenExponentN(6)
-                .mul(new Dec(amount))
-                .truncate()
-                .toString(),
-            },
-          ],
-        }).finish(),
-      };
-
-      try {
-        const gasUsed = await simulateMsgs(
-          OsmosisChainInfo,
-          key.bech32Address,
-          [protoMsgs],
-          [{ denom: "uosmo", amount: "236" }]
-        );
-
-        if (gasUsed) {
-          await sendMsgs(
-            window.keplr,
-            OsmosisChainInfo,
-            key.bech32Address,
-            [protoMsgs],
-            {
-              amount: [{ denom: "uosmo", amount: "236" }],
-              gas: Math.floor(gasUsed * 1.5).toString(),
-            }
-          );
-        }
-      } catch (e) {
-        if (e instanceof Error) {
-          console.log(e.message);
-        }
-      }
-    }
-  };
+  const [activeTab, setActiveTab] = React.useState<TabType>("osmosis");
 
   return (
     <div className="root-container">
@@ -156,545 +29,37 @@ function App() {
         />
       </div>
 
-      <h2 style={{ marginTop: "30px" }}>
-        Request to Osmosis Testnet via Keplr Provider
-      </h2>
-
-      <div className="item-container">
-        <div className="item">
-          <div className="item-title">Get OSMO Address</div>
-
-          <div className="item-content">
-            <div>
-              <button className="keplr-button" onClick={getKeyFromKeplr}>
-                Get Address
-              </button>
-            </div>
-            <div>Address: {address}</div>
-          </div>
-        </div>
-
-        <div className="item">
-          <div className="item-title">Get OSMO Balance</div>
-
-          <div className="item-content">
-            <button className="keplr-button" onClick={getBalance}>
-              Get Balance
-            </button>
-
-            <div>Balance: {balance}</div>
-          </div>
-        </div>
-
-        <div className="item">
-          <div className="item-title">Send OSMO</div>
-
-          <div className="item-content">
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "column",
-              }}
-            >
-              Recipient:
-              <input
-                type="text"
-                value={recipient}
-                onChange={(e) => setRecipient(e.target.value)}
-              />
-            </div>
-
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "column",
-              }}
-            >
-              Amount:
-              <input
-                type="text"
-                value={amount}
-                onChange={(e) => setAmount(e.target.value)}
-              />
-            </div>
-
-            <button className="keplr-button" onClick={sendBalance}>
-              Send
-            </button>
-          </div>
-        </div>
-      </div>
-
-      <h2 style={{ marginTop: "60px" }}>
-        Request to EVM Chain via Keplr Ethereum Provider
-      </h2>
-
       <div
-        className="item-container"
-        style={{ maxWidth: 576, overflowWrap: "anywhere" }}
+        style={{
+          display: "flex",
+          justifyContent: "center",
+          gap: "10px",
+          margin: "16px 0",
+        }}
       >
-        {keplrEip6963ProviderInfo && (
-          <div className="item">
-            <div className="item-title">EIP-6963 Provider Info</div>
-
-            <div className="item-content">
-              <div>UUID: {keplrEip6963ProviderInfo.uuid}</div>
-              <div>Name: {keplrEip6963ProviderInfo.name}</div>
-              <div>RDNS: {keplrEip6963ProviderInfo.rdns}</div>
-              <div>
-                Icon:{" "}
-                <img
-                  src={keplrEip6963ProviderInfo.icon}
-                  alt={keplrEip6963ProviderInfo.name}
-                  width={32}
-                  height={32}
-                />
-              </div>
-            </div>
-          </div>
-        )}
-
-        <div className="item">
-          <div className="item-title">Get Data From Wallet</div>
-
-          <div className="item-content">
-            {[
-              {
-                method: "eth_chainId",
-              },
-              {
-                method: "eth_accounts",
-              },
-            ].map(({ method }) => (
-              <div key={method}>
-                <button
-                  className="keplr-button"
-                  onClick={async () => {
-                    const result = await window.keplr?.ethereum?.request({
-                      method,
-                    });
-                    setEvmResult1(result.toString());
-                  }}
-                >
-                  {method}
-                </button>
-              </div>
-            ))}
-
-            <div>Result: {evmResult1}</div>
-          </div>
-        </div>
-
-        <div className="item">
-          <div className="item-title">Get Data From RPC Node</div>
-
-          <div className="item-content">
-            {[
-              {
-                method: "eth_blockNumber",
-              },
-              {
-                method: "eth_getBalance",
-              },
-            ].map(({ method }) => (
-              <div key={method}>
-                <button
-                  className="keplr-button"
-                  onClick={async () => {
-                    const params = (() => {
-                      switch (method) {
-                        case "eth_getBalance":
-                          return [
-                            window.keplr?.ethereum?.selectedAddress,
-                            "latest",
-                          ];
-                        default:
-                          return;
-                      }
-                    })();
-
-                    const result = await window.keplr?.ethereum?.request({
-                      method,
-                      params,
-                    });
-                    setEvmResult2(result);
-                  }}
-                >
-                  {method}
-                </button>
-              </div>
-            ))}
-
-            <div>Result: {evmResult2}</div>
-          </div>
-        </div>
-
-        <div className="item">
-          <div className="item-title">Sign Data</div>
-
-          <div className="item-content">
-            {[
-              {
-                method: "personal_sign",
-              },
-              {
-                method: "eth_signTypedData_v3",
-              },
-              {
-                method: "eth_signTypedData_v4",
-              },
-            ].map(({ method }) => (
-              <div key={method}>
-                <button
-                  className="keplr-button"
-                  onClick={async () => {
-                    const params = (() => {
-                      switch (method) {
-                        case "personal_sign":
-                          return [
-                            "0x536d9d5261206b3fd7ca989fe1c0afa69de8a1c673a14fcd5c98e1a641fad979",
-                            window.keplr?.ethereum?.selectedAddress,
-                          ];
-                        case "eth_signTypedData_v3":
-                          return [
-                            window.keplr?.ethereum?.selectedAddress,
-                            {
-                              types: {
-                                EIP712Domain: [
-                                  {
-                                    name: "name",
-                                    type: "string",
-                                  },
-                                  {
-                                    name: "version",
-                                    type: "string",
-                                  },
-                                  {
-                                    name: "chainId",
-                                    type: "uint256",
-                                  },
-                                  {
-                                    name: "verifyingContract",
-                                    type: "address",
-                                  },
-                                ],
-                                Person: [
-                                  {
-                                    name: "name",
-                                    type: "string",
-                                  },
-                                  {
-                                    name: "wallet",
-                                    type: "address",
-                                  },
-                                ],
-                                Mail: [
-                                  {
-                                    name: "from",
-                                    type: "Person",
-                                  },
-                                  {
-                                    name: "to",
-                                    type: "Person",
-                                  },
-                                  {
-                                    name: "contents",
-                                    type: "string",
-                                  },
-                                ],
-                              },
-                              primaryType: "Mail",
-                              domain: {
-                                name: "Ether Mail",
-                                version: 1,
-                                chainId: 1,
-                                verifyingContract:
-                                  "0xCcCCccccCCCCcCCCCCCcCcCccCcCCCcCcccccccC",
-                              },
-                              message: {
-                                from: {
-                                  name: "Cow",
-                                  wallet:
-                                    "0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826",
-                                },
-                                to: {
-                                  name: "Bob",
-                                  wallet:
-                                    "0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB",
-                                },
-                                contents: "Hello, Bob!",
-                              },
-                            },
-                          ];
-                        case "eth_signTypedData_v4":
-                          return [
-                            window.keplr?.ethereum?.selectedAddress,
-                            {
-                              domain: {
-                                chainId: 1,
-                                name: "Ether Mail",
-                                verifyingContract:
-                                  "0xCcCCccccCCCCcCCCCCCcCcCccCcCCCcCcccccccC",
-                                version: "1",
-                              },
-                              message: {
-                                contents: "Hello, Bob!",
-                                attachedMoneyInEth: 4.2,
-                                from: {
-                                  name: "Cow",
-                                  wallets: [
-                                    "0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826",
-                                    "0xDeaDbeefdEAdbeefdEadbEEFdeadbeEFdEaDbeeF",
-                                  ],
-                                },
-                                to: [
-                                  {
-                                    name: "Bob",
-                                    wallets: [
-                                      "0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB",
-                                      "0xB0BdaBea57B0BDABeA57b0bdABEA57b0BDabEa57",
-                                      "0xB0B0b0b0b0b0B000000000000000000000000000",
-                                    ],
-                                  },
-                                ],
-                              },
-                              // This refers to the keys of the following types object.
-                              primaryType: "Mail",
-                              types: {
-                                EIP712Domain: [
-                                  { name: "name", type: "string" },
-                                  { name: "version", type: "string" },
-                                  { name: "chainId", type: "uint256" },
-                                  {
-                                    name: "verifyingContract",
-                                    type: "address",
-                                  },
-                                ],
-                                Mail: [
-                                  { name: "from", type: "Person" },
-                                  { name: "to", type: "Person[]" },
-                                  { name: "contents", type: "string" },
-                                ],
-                                Person: [
-                                  { name: "name", type: "string" },
-                                  { name: "wallets", type: "address[]" },
-                                ],
-                              },
-                            },
-                          ];
-                        default:
-                          return [];
-                      }
-                    })();
-
-                    const result = await window.keplr?.ethereum?.request({
-                      method,
-                      params,
-                    });
-                    setEvmResult3(result);
-                  }}
-                >
-                  {method}
-                </button>
-              </div>
-            ))}
-
-            <div>Signature: {evmResult3}</div>
-          </div>
-        </div>
-
-        <div className="item">
-          <div className="item-title">Send Native Token</div>
-
-          <div className="item-content">
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "column",
-              }}
-            >
-              Recipient:
-              <input
-                type="text"
-                value={evmRecipient}
-                onChange={(e) => setEvmRecipient(e.target.value)}
-              />
-            </div>
-
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "column",
-              }}
-            >
-              Amount:
-              <input
-                type="number"
-                value={evmAmount}
-                onChange={(e) => setEvmAmount(e.target.value)}
-              />
-            </div>
-            <button
-              className="keplr-button"
-              onClick={async () => {
-                const parsedEvmAmount = parseFloat(evmAmount);
-
-                if (isNaN(parsedEvmAmount) || !isAddress(evmRecipient)) {
-                  return;
-                }
-
-                const currentChainId = window.keplr?.ethereum?._currentChainId;
-
-                const currentChainInfo =
-                  await window.keplr?.getChainInfoWithoutEndpoints(
-                    currentChainId
-                  );
-
-                if (currentChainInfo === undefined) {
-                  return;
-                }
-
-                const mainCurrency =
-                  currentChainInfo.stakeCurrency ??
-                  currentChainInfo.currencies[0];
-
-                const amountValue = `0x${parseInt(
-                  new Dec(parsedEvmAmount)
-                    .mulTruncate(
-                      DecUtils.getTenExponentN(mainCurrency.coinDecimals)
-                    )
-                    .toString()
-                ).toString(16)}`;
-
-                const tx = {
-                  from: window.keplr?.ethereum?.selectedAddress,
-                  to: evmRecipient,
-                  value: amountValue,
-                };
-
-                const result = await window.keplr?.ethereum?.request({
-                  method: "eth_sendTransaction",
-                  params: [tx],
-                });
-                setEvmResult4(result);
-              }}
-            >
-              {"eth_sendTransaction"}
-            </button>
-
-            <div>Transaction Hash: {evmResult4}</div>
-          </div>
-        </div>
-
-        <div className="item">
-          <div className="item-title">Add an EVM Chain</div>
-
-          <div className="item-content">
-            <div>Chain ID: 0x64</div>
-            <div>ChainName: Gnosis</div>
-            <div>RPC URL: https://rpc.gnosischain.com</div>
-            <button
-              className="keplr-button"
-              onClick={async () => {
-                await window.keplr?.ethereum?.request({
-                  method: "wallet_addEthereumChain",
-                  params: [
-                    {
-                      chainId: "0x64",
-                      chainName: "Gnosis",
-                      rpcUrls: [
-                        "https://rpc.gnosischain.com",
-                        "wss://rpc.gnosischain.com",
-                      ],
-                      iconUrls: [
-                        "https://xdaichain.com/fake/example/url/xdai.svg",
-                        "https://xdaichain.com/fake/example/url/xdai.png",
-                      ],
-                      nativeCurrency: {
-                        name: "XDAI",
-                        symbol: "XDAI",
-                        decimals: 18,
-                      },
-                      blockExplorerUrls: ["https://blockscout.com/poa/xdai/"],
-                    },
-                  ],
-                });
-              }}
-            >
-              {"wallet_addEthereumChain"}
-            </button>
-          </div>
-        </div>
-        <div className="item">
-          <div className="item-title">Switch to another EVM Chain</div>
-
-          <div className="item-content">
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "column",
-              }}
-            >
-              EVM Chain ID:
-              <input
-                type="number"
-                value={evmChainId}
-                onChange={(e) => setEvmChainId(e.target.value)}
-              />
-            </div>
-            <button
-              className="keplr-button"
-              onClick={async () => {
-                await window.keplr?.ethereum?.request({
-                  method: "wallet_switchEthereumChain",
-                  params: [
-                    {
-                      chainId: `0x${parseInt(evmChainId).toString(16)}`,
-                    },
-                  ],
-                });
-              }}
-            >
-              {"wallet_switchEthereumChain"}
-            </button>
-          </div>
-        </div>
-        <div className="item">
-          <div className="item-title">Add ERC20 Token</div>
-
-          <div className="item-content">
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "column",
-              }}
-            >
-              ERC20 Token Address:
-              <input
-                type="text"
-                value={evmTokenAddress}
-                onChange={(e) => setEvmTokenAddress(e.target.value)}
-              />
-            </div>
-            <button
-              className="keplr-button"
-              onClick={async () => {
-                await window.keplr?.ethereum?.request({
-                  method: "wallet_watchAsset",
-                  params: {
-                    type: "ERC20",
-                    options: {
-                      address: evmTokenAddress,
-                    },
-                  },
-                });
-              }}
-            >
-              {"wallet_watchAsset"}
-            </button>
-          </div>
-        </div>
+        {[
+          { id: "osmosis", label: "Osmosis" },
+          { id: "evm", label: "EVM" },
+          { id: "starknet", label: "Starknet" },
+          { id: "bitcoin", label: "Bitcoin" },
+        ].map((tab) => (
+          <button
+            key={tab.id}
+            className="keplr-button"
+            onClick={() => setActiveTab(tab.id as TabType)}
+            style={{
+              backgroundColor: activeTab === tab.id ? undefined : "#3C3C3C",
+              minWidth: "120px",
+            }}
+          >
+            {tab.label}
+          </button>
+        ))}
       </div>
+      {activeTab === "osmosis" && <OsmosisTab />}
+      {activeTab === "evm" && <EvmTab />}
+      {activeTab === "starknet" && <StarknetTab />}
+      {activeTab === "bitcoin" && <BitcoinTab />}
     </div>
   );
 }
