+++ src/components/BitcoinTab.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+export const BitcoinTab: React.FC = () => {
+  return (
+    <div className="item-container">
+      <div className="item">
+        <div className="item-title">Bitcoin Integration</div>
+        <div className="item-content">
+          <p>Bitcoin integration features will be implemented here.</p>
+          <p>Coming soon...</p>
+        </div>
+      </div>
+    </div>
+  );
+};
