+++ src/components/OsmosisTab.tsx
@@ -1,0 +1,186 @@
+import React, { useEffect, useState } from "react";
+import { Dec, DecUtils } from "@keplr-wallet/unit";
+import { OsmosisChainInfo } from "../constants";
+import { Balances } from "../types/balance";
+import { api } from "../util/api";
+import { sendMsgs } from "../util/sendMsgs";
+import { simulateMsgs } from "../util/simulateMsgs";
+import { MsgSend } from "../proto-types-gen/src/cosmos/bank/v1beta1/tx";
+
+export const OsmosisTab: React.FC = () => {
+  const [address, setAddress] = useState<string>("");
+  const [balance, setBalance] = useState<string>("");
+  const [recipient, setRecipient] = useState<string>("");
+  const [amount, setAmount] = useState<string>("");
+
+  const init = async () => {
+    const keplr = window.keplr;
+    if (keplr) {
+      try {
+        await keplr.experimentalSuggestChain(OsmosisChainInfo);
+        if (!keplr.ethereum?.isConnected()) {
+          await keplr.ethereum?.enable();
+        }
+      } catch (e) {
+        if (e instanceof Error) {
+          console.log(e.message);
+        }
+      }
+    }
+  };
+
+  const getKeyFromKeplr = async () => {
+    const key = await window.keplr?.getKey(OsmosisChainInfo.chainId);
+    if (key) {
+      setAddress(key.bech32Address);
+    }
+  };
+
+  const getBalance = async () => {
+    const key = await window.keplr?.getKey(OsmosisChainInfo.chainId);
+
+    if (key) {
+      const uri = `${OsmosisChainInfo.rest}/cosmos/bank/v1beta1/balances/${key.bech32Address}?pagination.limit=1000`;
+
+      const data = await api<Balances>(uri);
+      const balance = data.balances.find(
+        (balance) => balance.denom === "uosmo"
+      );
+      const osmoDecimal = OsmosisChainInfo.currencies.find(
+        (currency) => currency.coinMinimalDenom === "uosmo"
+      )?.coinDecimals;
+
+      if (balance) {
+        const amount = new Dec(balance.amount, osmoDecimal);
+        setBalance(`${amount.toString(osmoDecimal)} OSMO`);
+      } else {
+        setBalance(`0 OSMO`);
+      }
+    }
+  };
+
+  const sendBalance = async () => {
+    if (window.keplr) {
+      const key = await window.keplr?.getKey(OsmosisChainInfo.chainId);
+      const protoMsgs = {
+        typeUrl: "/cosmos.bank.v1beta1.MsgSend",
+        value: MsgSend.encode({
+          fromAddress: key.bech32Address,
+          toAddress: recipient,
+          amount: [
+            {
+              denom: "uosmo",
+              amount: DecUtils.getTenExponentN(6)
+                .mul(new Dec(amount))
+                .truncate()
+                .toString(),
+            },
+          ],
+        }).finish(),
+      };
+
+      try {
+        const gasUsed = await simulateMsgs(
+          OsmosisChainInfo,
+          key.bech32Address,
+          [protoMsgs],
+          [{ denom: "uosmo", amount: "236" }]
+        );
+
+        if (gasUsed) {
+          await sendMsgs(
+            window.keplr,
+            OsmosisChainInfo,
+            key.bech32Address,
+            [protoMsgs],
+            {
+              amount: [{ denom: "uosmo", amount: "236" }],
+              gas: Math.floor(gasUsed * 1.5).toString(),
+            }
+          );
+        }
+      } catch (e) {
+        if (e instanceof Error) {
+          console.log(e.message);
+        }
+      }
+    }
+  };
+
+  useEffect(() => {
+    init();
+  }, []);
+
+  return (
+    <>
+      <h2 style={{ marginTop: "30px" }}>
+        Request to Osmosis Testnet via Keplr Provider
+      </h2>
+
+      <div className="item-container">
+        <div className="item">
+          <div className="item-title">Get OSMO Address</div>
+
+          <div className="item-content">
+            <div>
+              <button className="keplr-button" onClick={getKeyFromKeplr}>
+                Get Address
+              </button>
+            </div>
+            <div>Address: {address}</div>
+          </div>
+        </div>
+
+        <div className="item">
+          <div className="item-title">Get OSMO Balance</div>
+
+          <div className="item-content">
+            <button className="keplr-button" onClick={getBalance}>
+              Get Balance
+            </button>
+
+            <div>Balance: {balance}</div>
+          </div>
+        </div>
+
+        <div className="item">
+          <div className="item-title">Send OSMO</div>
+
+          <div className="item-content">
+            <div
+              style={{
+                display: "flex",
+                flexDirection: "column",
+              }}
+            >
+              Recipient:
+              <input
+                type="text"
+                value={recipient}
+                onChange={(e) => setRecipient(e.target.value)}
+              />
+            </div>
+
+            <div
+              style={{
+                display: "flex",
+                flexDirection: "column",
+              }}
+            >
+              Amount:
+              <input
+                type="text"
+                value={amount}
+                onChange={(e) => setAmount(e.target.value)}
+              />
+            </div>
+
+            <button className="keplr-button" onClick={sendBalance}>
+              Send
+            </button>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+};
