+++ src/components/StarknetTab.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+export const StarknetTab: React.FC = () => {
+  return (
+    <div className="item-container">
+      <div className="item">
+        <div className="item-title">Starknet Integration</div>
+        <div className="item-content">
+          <p>Starknet integration features will be implemented here.</p>
+          <p>Coming soon...</p>
+        </div>
+      </div>
+    </div>
+  );
+};
